+++ test/unit/AsThis.cs
@@ -127,7 +127,7 @@ namespace DecTest
             var item = new ThisThenClassOuter();
             item.data = new ThisThenClassInnerDerived();
 
-            var deserialized = DoRecorderRoundTrip(item, mode, expectWriteErrors: mode != RecorderMode.Clone);
+            var deserialized = DoRecorderRoundTrip(item, mode, expectWriteErrors: mode != RecorderMode.Clone && mode != RecorderMode.Checksum);
         }
 
         public class ClassThenThisOuterBase : Dec.IRecordable
@@ -147,7 +147,7 @@ namespace DecTest
             var item = new ClassThenThisOuterDerived();
             ClassThenThisOuterBase itemBase = item;
 
-            var deserialized = DoRecorderRoundTrip(itemBase, mode, expectReadErrors: mode != RecorderMode.Clone, expectWriteErrors: mode != RecorderMode.Clone);
+            var deserialized = DoRecorderRoundTrip(itemBase, mode, expectReadErrors: mode != RecorderMode.Clone && mode != RecorderMode.Checksum, expectWriteErrors: mode != RecorderMode.Clone && mode != RecorderMode.Checksum);
         }
 
         public class Inner : Dec.IRecordable